Move thrift classes to different package
