+++ jaeger-crossdock/src/main/java/com/uber/jaeger/crossdock/tracetest_manual/TraceResponse.java
@@ -78,28 +78,28 @@ public class TraceResponse {
                 '}';
     }
 
-    public static com.uber.jaeger.crossdock.tracetest.TraceResponse toThrift(TraceResponse resp) {
+    public static com.uber.jaeger.crossdock.thrift.TraceResponse toThrift(TraceResponse resp) {
         if (resp == null) {
             return null;
         }
-        com.uber.jaeger.crossdock.tracetest.TraceResponse response =
-                new com.uber.jaeger.crossdock.tracetest.TraceResponse();
+        com.uber.jaeger.crossdock.thrift.TraceResponse response =
+                new com.uber.jaeger.crossdock.thrift.TraceResponse();
         response.setSpan(toThrift(resp.getObservedSpan()));
         response.setDownstream(toThrift(resp.getDownstream()));
         response.setNotImplementedError("");
         return response;
     }
 
-    private static com.uber.jaeger.crossdock.tracetest.ObservedSpan toThrift(ObservedSpan observedSpan) {
-        com.uber.jaeger.crossdock.tracetest.ObservedSpan res =
-                new com.uber.jaeger.crossdock.tracetest.ObservedSpan();
+    private static com.uber.jaeger.crossdock.thrift.ObservedSpan toThrift(ObservedSpan observedSpan) {
+        com.uber.jaeger.crossdock.thrift.ObservedSpan res =
+                new com.uber.jaeger.crossdock.thrift.ObservedSpan();
         res.setTraceId(observedSpan.getTraceID());
         res.setBaggage(observedSpan.getBaggage());
         res.setSampled(observedSpan.getSampled());
         return res;
     }
 
-    public static TraceResponse fromThrift(com.uber.jaeger.crossdock.tracetest.TraceResponse thrift) {
+    public static TraceResponse fromThrift(com.uber.jaeger.crossdock.thrift.TraceResponse thrift) {
         if (thrift == null) {
             return null;
         }
